# platform/apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "operationId": "ProductsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/products/barcode/{code}": {
      "get": {
        "operationId": "ProductsController_byBarcode",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/products/{id}": {
      "get": {
        "operationId": "ProductsController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "ProductsController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/products/{id}/restore": {
      "put": {
        "operationId": "ProductsController_restore",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/products/{id}/restock": {
      "post": {
        "operationId": "ProductsController_restock",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "UZANITE API",
    "description": "UZANITE platform API",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}